+++ packages/react-wrap-with/src/private/wrapWith.tsx
@@ -1,9 +1,9 @@
 import React, { createElement, forwardRef, Fragment } from 'react';
 
-import ExtractProp from '../ExtractProp';
+import Extract from '../ExtractProp';
 import pick from './util/pick';
 import pickAndOmit from './util/pickAndOmit';
-import SpyProp from '../SpyProp';
+import Spy from '../SpyProp';
 
 import type { ComponentType, PropsWithChildren, PropsWithoutRef, ReactNode, RefAttributes } from 'react';
 import type { HowOf } from '../HowOf';
@@ -16,9 +16,9 @@ const EmptyComponent = () => <Fragment />;
 export default function wrapWith<
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ContainerComponentType extends ComponentType<PropsWithChildren<any>>,
-  ExtractPropKeys extends keyof PropsOf<ContainerComponentType> = never,
-  SpyPropKeys extends keyof PropsOf<ContainerComponentType> = never,
-  const How extends HowOf<ContainerComponentType, ExtractPropKeys, SpyPropKeys> = HowOf<
+  ExtractPropKeys extends keyof ContainerComponentType,
+  SpyPropKeys extends keyof ContainerComponentType,
+  const How extends HowOf<PropsOf<ContainerComponentType>, ExtractPropKeys, SpyPropKeys> = HowOf<
     ContainerComponentType,
     ExtractPropKeys,
     SpyPropKeys
@@ -27,28 +27,22 @@ export default function wrapWith<
   ContainerComponent: ContainerComponentType | false | null | undefined,
   how: How
 ): <
-  ContentComponentType extends ComponentType<
-    Pick<
-      PropsOf<ContainerComponentType>,
-      keyof {
-        [K in keyof typeof how as (typeof how)[K] extends typeof SpyProp ? K : never]: (typeof how)[K];
-      }
-    >
+  MinimalContentProps extends Pick<
+    PropsOf<ContainerComponentType>,
+    keyof {
+      [K in keyof How as How[K] extends typeof Extract ? K : How[K] extends typeof Spy ? K : never]: How[K];
+    }
   >,
-  Ref = RefOf<PropsOf<ContentComponentType>>
+  Ref = RefOf<MinimalContentProps>
 >(
-  ContentComponent: ContentComponentType | false | null | undefined
+  contentComponent: ComponentType<MinimalContentProps> | false | null | undefined
 ) => ComponentType<
   PropsWithoutRef<
-    PropsOf<ContentComponentType> &
+    PropsOf<typeof contentComponent> &
       Pick<
         PropsOf<ContainerComponentType>,
         keyof {
-          [K in keyof typeof how as (typeof how)[K] extends typeof ExtractProp
-            ? K
-            : (typeof how)[K] extends typeof SpyProp
-            ? K
-            : never]: (typeof how)[K];
+          [K in keyof How as How[K] extends typeof Extract ? K : never]: How[K];
         }
       >
   > &
@@ -57,8 +51,7 @@ export default function wrapWith<
 
 // If <Container> need no props other than children, initialProps is optional.
 export default function wrapWith<
-  ContainerComponentType extends // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ComponentType<{ children?: ReactNode | undefined }> | false | null | undefined
+  ContainerComponentType extends ComponentType<{ children?: ReactNode | undefined }> | false | null | undefined
 >(
   ContainerComponent: ContainerComponentType | false | null | undefined,
   how?: undefined | Record<number | string | symbol, never>
@@ -79,53 +72,47 @@ export default function wrapWith<
   >
 >(ContainerComponent: ContainerComponentType | false | null | undefined, how: How) {
   type ContainerProps = PropsOf<ContainerComponentType>;
-
   type ExtractPropsKeys = keyof {
-    [K in keyof How as How[K] extends typeof ExtractProp ? K : never]: How[K];
+    [K in keyof How as How[K] extends typeof Extract ? K : never]: How[K];
   };
   type SpyPropsKeys = keyof {
-    [K in keyof How as How[K] extends typeof SpyProp ? K : never]: How[K];
+    [K in keyof How as How[K] extends typeof Spy ? K : never]: How[K];
   };
   type InitialPropsKeys = keyof {
-    [K in keyof How as How[K] extends typeof ExtractProp ? never : How[K] extends typeof SpyProp ? never : K]: How[K];
+    [K in keyof How as How[K] extends typeof Extract ? never : How[K] extends typeof Spy ? never : K]: How[K];
   };
 
   type ExtractProps = Pick<ContainerProps, ExtractPropsKeys>;
   type SpyProps = Pick<ContainerProps, SpyPropsKeys>;
   type InitialProps = Omit<ContainerProps, ExtractPropsKeys | SpyPropsKeys>;
+  // Could we simplified?
+  // type ExtractProps = { [K in keyof How as How[K] extends typeof Extract ? K : never]: ContainerProps[K] };
+  // type SpyProps = { [K in keyof How as How[K] extends typeof Spy ? K : never]: ContainerProps[K] };
+  // type InitialProps = Omit<ContainerProps, keyof ExtractProps | keyof SpyProps>;
 
   const extractPropsKeys: ExtractPropsKeys[] = Object.entries(how || {})
-    .filter(([_, value]) => value === ExtractProp)
+    .filter(([_, value]) => value === Extract)
     .map(([key]) => key as keyof How) as ExtractPropsKeys[];
 
   const spyPropsKeys: SpyPropsKeys[] = Object.entries(how || {})
-    .filter(([_, value]) => value === SpyProp)
+    .filter(([_, value]) => value === Spy)
     .map(([key]) => key as keyof How) as SpyPropsKeys[];
 
   const initialPropsKeys: InitialPropsKeys[] = Object.entries(how || {})
-    .filter(([_, value]) => value !== ExtractProp && value !== SpyProp)
+    .filter(([_, value]) => value !== Extract && value !== Spy)
     .map(([key]) => key as keyof How) as InitialPropsKeys[];
 
   // Try fix this.
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const [initialProps] = pickAndOmit<InitialProps, ExtractProps | SpyProps>(how as any, initialPropsKeys as any);
 
-  return function wrap<
-    ContentComponentType extends ComponentType<ContentProps>,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ContentProps extends Pick<ContainerProps, SpyPropsKeys> = Pick<ContainerProps, SpyPropsKeys>,
-    Ref = RefOf<ContentProps>
-  >(
-    ContentComponent: // | { how: How; extract: ExtractPropsKeys; spy: SpyPropsKeys; initial: InitialPropsKeys }
-    ContentComponentType | false | null | undefined
-  ): ComponentType<PropsWithoutRef<PropsOf<ContentComponentType> & ExtractProps & SpyProps> & RefAttributes<Ref>> {
+  return function wrap<MinimalContentProps extends SpyProps & ExtractProps, Ref = RefOf<MinimalContentProps>>(
+    contentComponent: ComponentType<MinimalContentProps> | false | null | undefined
+  ): ComponentType<PropsWithoutRef<PropsOf<typeof contentComponent> & ExtractProps> & RefAttributes<Ref>> {
     if (ContainerComponent) {
-      const WithContainer = forwardRef<Ref, PropsOf<ContentComponentType> & ExtractProps>((props, ref) => {
-        const [extractedProps, contentProps] = pickAndOmit<ExtractProps, PropsOf<ContentComponentType>>(
-          props,
-          extractPropsKeys
-        );
-        const spyProps = pick(props, spyPropsKeys);
+      const WithContainer = forwardRef<Ref, MinimalContentProps>((props, ref) => {
+        const [extractedProps, contentProps] = pickAndOmit<ExtractProps, MinimalContentProps>(props, extractPropsKeys);
+        const spyProps = pick<MinimalContentProps, SpyPropsKeys>(props, spyPropsKeys);
 
         return createElement(
           ContainerComponent,
@@ -133,7 +120,7 @@ export default function wrapWith<
           // If there are "ContentComponentType" is falsy, don't override children. It will override the `props.children`.
           // False positive: we are unpacking the array rightaway.
           // eslint-disable-next-line react/jsx-key
-          ...(ContentComponent ? [createElement(ContentComponent, { ...contentProps, ref })] : [])
+          ...(contentComponent ? [createElement(contentComponent, { ...contentProps, ref })] : [])
         );
       });
 
@@ -142,20 +129,17 @@ export default function wrapWith<
       return WithContainer;
     }
 
-    if (ContentComponent) {
-      const WithContainer = forwardRef<
-        Ref,
-        Pick<PropsOf<ContainerComponentType>, ExtractPropsKeys> & PropsOf<ContentComponentType>
-      >((props, ref) => {
+    if (contentComponent) {
+      const WithContainer = forwardRef<Ref, MinimalContentProps>((props, ref) => {
         const [, contentProps] = pickAndOmit<
           Pick<PropsOf<ContainerComponentType>, ExtractPropsKeys>,
-          PropsOf<ContentComponentType> & { ref?: Ref }
+          MinimalContentProps
         >(props, extractPropsKeys);
 
-        return createElement(ContentComponent, { ...contentProps, ref });
+        return createElement(contentComponent, { ...contentProps, ref });
       });
 
-      WithContainer.displayName = ContentComponent.displayName;
+      WithContainer.displayName = contentComponent.displayName;
 
       return WithContainer;
     }